feat(department):  + Update URLs.  + Display pages based on user permissions.  + Create account information edit page for users.  + Create API to allow users to edit their account details.  + Create API to change password.  + Update API for creating and updating bank accounts.  + Create API to fetch bank account details by ID.

+++ src/main/java/com/tytngn/fundsmanagement/service/UserService.java
@@ -1,13 +1,16 @@
 package com.tytngn.fundsmanagement.service;
 
-import com.tytngn.fundsmanagement.dto.request.UserCreationRequest;
-import com.tytngn.fundsmanagement.dto.request.UserUpdateRequest;
+import com.tytngn.fundsmanagement.configuration.SecurityExpression;
+import com.tytngn.fundsmanagement.dto.request.*;
 import com.tytngn.fundsmanagement.dto.response.UserResponse;
+import com.tytngn.fundsmanagement.entity.BankAccount;
 import com.tytngn.fundsmanagement.entity.Role;
 import com.tytngn.fundsmanagement.entity.User;
 import com.tytngn.fundsmanagement.exception.AppException;
 import com.tytngn.fundsmanagement.exception.ErrorCode;
+import com.tytngn.fundsmanagement.mapper.BankAccountMapper;
 import com.tytngn.fundsmanagement.mapper.UserMapper;
+import com.tytngn.fundsmanagement.repository.BankAccountRepository;
 import com.tytngn.fundsmanagement.repository.DepartmentRepository;
 import com.tytngn.fundsmanagement.repository.RoleRepository;
 import com.tytngn.fundsmanagement.repository.UserRepository;
@@ -15,14 +18,13 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -32,10 +34,15 @@ import java.util.stream.Collectors;
 public class UserService {
 
     UserRepository userRepository;
-    UserMapper userMapper;
     RoleRepository roleRepository;
-    PasswordEncoder passwordEncoder;
     DepartmentRepository departmentRepository;
+    BankAccountRepository bankAccountRepository;
+
+    UserMapper userMapper;
+    BankAccountMapper bankAccountMapper;
+
+    PasswordEncoder passwordEncoder;
+    SecurityExpression securityExpression;
 
     // tạo User
     public UserResponse createUser(UserCreationRequest request) {
@@ -88,11 +95,17 @@ public class UserService {
 
     // lấy thông tin user đang đăng nhập
     public UserResponse getMyInfo(){
-        var context = SecurityContextHolder.getContext();
-        String name = context.getAuthentication().getName();
+        // Lấy thông tin người dùng đang đăng nhập
+        String id = securityExpression.getUserId();
 
-        User user = userRepository.findByUsername(name).orElseThrow(() ->
-                new AppException(ErrorCode.USER_NOT_EXISTS));
+        // Tìm kiếm người dùng trong cơ sở dữ liệu
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+
+        if(user.getStatus() == 9999){
+            Set<Role> roles = new HashSet<>(roleRepository.findAll());
+            user.setRoles(roles);
+        }
 
         return userMapper.toUserResponse(user);
     }
@@ -123,7 +136,7 @@ public class UserService {
     }
 
 
-    // update User dựa trên id
+    // Admin update User dựa trên id
     public UserResponse updateUser(String userId, UserUpdateRequest request) {
 
         // Kiểm tra user có tồn tại không
@@ -132,6 +145,9 @@ public class UserService {
 
         userMapper.updateUser(user, request);
 //        user.setPassword(passwordEncoder.encode(request.getPassword()));
+        if (user.getStatus() == 9999){
+            throw new AppException(ErrorCode.ADMIN_NOT_EDITABLE);
+        }
 
         var roles = roleRepository.findAllById(request.getRoleId());
         user.setRoles(new HashSet<>(roles));
@@ -143,6 +159,50 @@ public class UserService {
         user.setUpdatedDate(LocalDate.now());
 
         return userMapper.toUserResponse(userRepository.save(user));
+    }
+
+
+    // Người dùng update tài khoản bao gồm tên ngân hàng và số tài khoản
+    public UserResponse updateUserAccount(AccountUpdateRequest request) {
+        // Lấy thông tin người dùng hiện tại từ token
+        String userId = securityExpression.getUserId();
+
+        // Tìm kiếm người dùng trong cơ sở dữ liệu
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+
+        // Không cho phép người dùng chỉnh sửa nếu trạng thái tài khoản là 9999
+        if (user.getStatus() == 9999) {
+            throw new AppException(ErrorCode.ADMIN_NOT_EDITABLE);
+        }
+
+        // Cập nhật thông tin cá nhân
+        userMapper.updateUserAccount(user, request);
+
+        // Cập nhật hoặc tạo mới tài khoản ngân hàng
+        BankAccount bankAccount = user.getAccount();
+        BankAccountRequest bankAccountRequest = BankAccountRequest.builder()
+                .bankName(request.getBankName())
+                .accountNumber(request.getAccountNumber())
+                .build();
+        if (bankAccount == null) {
+            bankAccount = bankAccountMapper.toBankAccount(bankAccountRequest);
+            bankAccount.setUser(user);
+            bankAccount.setCreatedDate(LocalDate.now());
+            bankAccountRepository.save(bankAccount);
+        } else {
+            bankAccountMapper.updateBankAccounts(bankAccount, request);
+            bankAccount.setCreatedDate(LocalDate.now());
+            bankAccountRepository.save(bankAccount);
+        }
+
+        user.setAccount(bankAccount);
+        user.setUpdatedDate(LocalDate.now());
+
+        // Lưu thay đổi
+        userRepository.save(user);
+
+        return userMapper.toUserResponse(user);
     }
 
 
@@ -163,12 +223,50 @@ public class UserService {
     }
 
 
+    // Đổi mật khẩu cho tài khoản
+    public UserResponse changePassword(ChangePasswordRequest request) {
+
+        // Lấy thông tin người dùng đang đăng nhập
+        String id = securityExpression.getUserId();
+
+        // Tìm kiếm người dùng trong cơ sở dữ liệu
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+
+        // Kiểm tra mật khẩu hiện tại có đúng không
+        if (!passwordEncoder.matches(request.getCurrentPassword(), user.getPassword())) {
+            throw new AppException(ErrorCode.CURRENT_PASSWORD_INCORRECT);
+        }
+
+        // Kiểm tra xem mật khẩu mới có khác mật khẩu hiện tại không
+        if (passwordEncoder.matches(request.getNewPassword(), user.getPassword())) {
+            throw new AppException(ErrorCode.NEW_PASSWORD_MUST_BE_DIFFERENT);
+        }
+
+        // Mã hóa mật khẩu mới và lưu lại
+        user.setPassword(passwordEncoder.encode(request.getNewPassword()));
+
+        // Cập nhật ngày thay đổi mật khẩu
+        user.setUpdatedDate(LocalDate.now());
+
+        // Lưu lại user
+        userRepository.save(user);
+
+        // Trả về UserResponse sau khi đổi mật khẩu
+        return userMapper.toUserResponse(user);
+    }
+
+
     // vô hiệu hóa tài khoản
     public UserResponse disableUser(String userId) {
 
         // Kiểm tra user có tồn tại không
         User user = userRepository.findById(userId).orElseThrow(() ->
                 new AppException(ErrorCode.USER_NOT_EXISTS));
+
+        if (user.getStatus() == 9999){
+            throw new AppException(ErrorCode.ADMIN_NOT_EDITABLE);
+        }
 
         // Cập nhật status = 0 để vô hiệu hoá tài khoản
         user.setStatus(0);
@@ -182,8 +280,12 @@ public class UserService {
 
 
     public void deleteUser(String userId) {
-        userRepository.findById(userId).orElseThrow(() ->
+        User user = userRepository.findById(userId).orElseThrow(() ->
                 new AppException(ErrorCode.USER_NOT_EXISTS));
+
+        if (user.getStatus() == 9999){
+            throw new AppException(ErrorCode.ADMIN_NOT_EDITABLE);
+        }
 
         userRepository.deleteById(userId);
     }
